+++ engine/utils.h
@@ -23,7 +23,7 @@ double clamp(double val, double lower, double upper)
   return min(max(val, lower), upper);
 }
 
-double find_peak(double* buffer, int buffer_len) {
+double find_peak(float* buffer, int buffer_len) {
   double peak = 0;
   for (int i = 0; i < buffer_len; ++i) {
     if (buffer[i] < -peak)
